mart_building_cv/requirements.txt
ultralytics
torch
torchvision
opencv-python
pyyaml
paho-mqtt